requirements.txt
asgiref==3.6.0
Django==4.1.5
django-crispy-forms==1.14.0
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
